requirements.txt
pandas
numpy
scikit-learn
joblib
requests
python-dotenv
google-generativeai
